fix a bug that writes to the wp-config.php file over and over again

+++ includes/classes/class-sbp-cache.php
@@ -331,20 +331,24 @@ class SBP_Cache extends SBP_Abstract_Module {
 		$advanced_cache_path = WP_CONTENT_DIR . '/advanced-cache.php';
 
 		if ( ! isset( $_SERVER['KINSTA_CACHE_ZONE'] ) && ( ! defined( 'IS_PRESSABLE' ) || ! IS_PRESSABLE ) ) {
-			// Delete or recreate advanced-cache.php
-			if ( $saved_data['module_caching'] ) {
-				$sbp_advanced_cache = SBP_PATH . '/advanced-cache.php';
 
-				SBP_Cache::set_wp_cache_constant( true );
+			if ( sbp_get_option( 'module_caching' ) !== $saved_data['module_caching'] ) {
 
-				file_put_contents( WP_CONTENT_DIR . '/advanced-cache.php', file_get_contents( $sbp_advanced_cache ) );
+				// Delete or recreate advanced-cache.php
+				if ( $saved_data['module_caching'] ) {
+					$sbp_advanced_cache = SBP_PATH . '/advanced-cache.php';
 
-				self::create_settings_json( $saved_data );
-			} else {
-				SBP_Cache::set_wp_cache_constant( false );
-				if ( file_exists( $advanced_cache_path ) ) {
-					if ( ! unlink( $advanced_cache_path ) ) {
-						return wp_send_json_error( [ 'notice' => esc_html__( 'advanced-cache.php can not be removed. Please remove it manually.', 'speed-booster-pack' ), 'errors' => [] ] );
+					SBP_Cache::set_wp_cache_constant( true );
+
+					file_put_contents( WP_CONTENT_DIR . '/advanced-cache.php', file_get_contents( $sbp_advanced_cache ) );
+
+					self::create_settings_json( $saved_data );
+				} else {
+					SBP_Cache::set_wp_cache_constant( false );
+					if ( file_exists( $advanced_cache_path ) ) {
+						if ( ! unlink( $advanced_cache_path ) ) {
+							return wp_send_json_error( [ 'notice' => esc_html__( 'advanced-cache.php can not be removed. Please remove it manually.', 'speed-booster-pack' ), 'errors' => [] ] );
+						}
 					}
 				}
 			}
